move more key/signature functionality from Chain to SectionKeysProvider
